feat(middleware-signing): separate and refactor sigV4 and awsAuth
