+++ packages/middleware-signing/package.json
@@ -24,6 +24,7 @@
     "@aws-sdk/protocol-http": "*",
     "@aws-sdk/signature-v4": "*",
     "@aws-sdk/types": "*",
+    "@aws-sdk/util-identity-auth": "*",
     "@aws-sdk/util-middleware": "*",
     "tslib": "^2.3.1"
   },
