--- packages/middleware-signing/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./configurations";
-export * from "./middleware";
