+++ packages/middleware-signing/src/resolveAwsAuthConfig.ts
@@ -1,0 +1,100 @@
+import { SignatureV4, SignatureV4CryptoInit, SignatureV4Init } from "@aws-sdk/signature-v4";
+import { AuthScheme, RegionInfo, RequestSigner } from "@aws-sdk/types";
+import { normalizeIdentityProvider } from "@aws-sdk/util-identity-auth";
+import { normalizeProvider } from "@aws-sdk/util-middleware";
+
+import { AwsAuthInputConfig, AwsAuthResolvedConfig, PreviouslyResolved } from "./configurations";
+
+export const resolveAwsAuthConfig = <T>(
+  input: T & AwsAuthInputConfig & PreviouslyResolved
+): T & AwsAuthResolvedConfig => {
+  const normalizedCreds = input.credentials
+    ? normalizeIdentityProvider(input.credentials)
+    : input.credentialDefaultProvider(input as any);
+  const { signingEscapePath = true, systemClockOffset = input.systemClockOffset || 0, sha256 } = input;
+  let signer: (authScheme?: AuthScheme) => Promise<RequestSigner>;
+  if (input.signer) {
+    // if signer is supplied by user, normalize it to a function returning a promise for signer.
+    signer = normalizeProvider(input.signer);
+  } else if (input.regionInfoProvider) {
+    // This branch is for endpoints V1.
+    // construct a provider inferring signing from region.
+    signer = () =>
+      normalizeProvider(input.region)()
+        .then(
+          async (region) =>
+            [
+              (await input.regionInfoProvider!(region, {
+                useFipsEndpoint: await input.useFipsEndpoint(),
+                useDualstackEndpoint: await input.useDualstackEndpoint(),
+              })) || {},
+              region,
+            ] as [RegionInfo, string]
+        )
+        .then(([regionInfo, region]) => {
+          const { signingRegion, signingService } = regionInfo;
+          // update client's singing region and signing service config if they are resolved.
+          // signing region resolving order: user supplied signingRegion -> endpoints.json inferred region -> client region
+          input.signingRegion = input.signingRegion || signingRegion || region;
+          // signing name resolving order:
+          // user supplied signingName -> endpoints.json inferred (credential scope -> model arnNamespace) -> model service id
+          input.signingName = input.signingName || signingService || input.serviceId;
+
+          const params: SignatureV4Init & SignatureV4CryptoInit = {
+            ...input,
+            credentials: normalizedCreds,
+            region: input.signingRegion,
+            service: input.signingName,
+            sha256,
+            uriEscapePath: signingEscapePath,
+          };
+          const SignerCtor = input.signerConstructor || SignatureV4;
+          return new SignerCtor(params);
+        });
+  } else {
+    // This branch is for endpoints V2.
+    // Handle endpoints v2 that resolved per-command
+    // TODO: need total refactor for reference auth architecture.
+    signer = async (authScheme?: AuthScheme) => {
+      authScheme = Object.assign(
+        {},
+        {
+          name: "sigv4",
+          signingName: input.signingName || input.defaultSigningName!,
+          signingRegion: await normalizeProvider(input.region)(),
+          properties: {},
+        },
+        authScheme
+      );
+
+      const signingRegion = authScheme.signingRegion;
+      const signingService = authScheme.signingName;
+      // update client's singing region and signing service config if they are resolved.
+      // signing region resolving order: user supplied signingRegion -> endpoints.json inferred region -> client region
+      input.signingRegion = input.signingRegion || signingRegion;
+      // signing name resolving order:
+      // user supplied signingName -> endpoints.json inferred (credential scope -> model arnNamespace) -> model service id
+      input.signingName = input.signingName || signingService || input.serviceId;
+
+      const params: SignatureV4Init & SignatureV4CryptoInit = {
+        ...input,
+        credentials: normalizedCreds,
+        region: input.signingRegion,
+        service: input.signingName,
+        sha256,
+        uriEscapePath: signingEscapePath,
+      };
+
+      const SignerCtor = input.signerConstructor || SignatureV4;
+      return new SignerCtor(params);
+    };
+  }
+
+  return {
+    ...input,
+    systemClockOffset,
+    signingEscapePath,
+    credentials: normalizedCreds,
+    signer,
+  };
+};
